+++ src/Config.hs
@@ -227,6 +227,7 @@ sourceFile ClashOfClans = "clash_of_clans"
 sourceFile Conan = "conan"
 sourceFile Doraemon = "doraemon"
 sourceFile FinalSpace = "final_space"
+sourceFile HowToTrainYourDragon = "how_to_train_your_dragon"
 
 mapSource :: Text -> Name
 mapSource "studioGhibli" = 'StudioGhibli
@@ -402,7 +403,8 @@ mapSource "camera" = 'Camera
 mapSource "clashOfClans" = 'ClashOfClans
 mapSource "conan" = 'Conan
 mapSource "doraemon" = 'Doraemon
-mapSource "final_space" = 'FinalSpace
+mapSource "finalSpace" = 'FinalSpace
+mapSource "howToTrainYourDragon" = 'HowToTrainYourDragon
 mapSource item = error $ "mapSource: Invalid argument passed " <> (show item)
 
 guessSourceFile :: SourceData -> Text -> FilePath